requirements.txt
apache-airflow==2.8.1
pyspark==3.5.0
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
sqlalchemy==1.4.51
python-dotenv==1.0.1
minio==7.2.3
pandas>=2.0.0
pytest==8.0.1
pytest-mock==3.12.0
wget==3.2
psutil==5.9.8
pendulum==2.1.2
py4j==0.10.9.7
numpy>=1.24.0
pyppeteer==1.0.2
websockets==10.3
nltk==3.8.1